+++ objc2-encode/CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased - YYYY-MM-DD
 
+### Changed
+* **BREAKING**: Renamed `Encoding::C_U_LONG` to `Encoding::C_ULONG`.
+
 
 ## 2.0.0-pre.0 - 2022-06-13